Fix warnings on clang. Fixed Document/Exe duo. CMake is fucking useless i can't build a fucking dll for 32 bit and 64 bit in a single CMakeFiles.

+++ src/Mouse.cpp
@@ -144,7 +144,7 @@ void MouseWindow::render(std::optional<MouseState>& state) noexcept {
 		ImGui::SameLine();
 
 		auto time_end = time(nullptr);
-		ImGui::Text("%d", Reset_Button_Time - (time_end - reset_time_start));
+		ImGui::Text("%d", (int)(Reset_Button_Time - (time_end - reset_time_start)));
 		if (time_end - reset_time_start + 1 > Reset_Button_Time) {
 			reset_time_start = 0;
 		}
